fix(pos): tolerant tax-code rate decode + walk-in customer fallback for on-account
- tax-codes 502: treasury serializes the tax rate as a STRING; pos-api's TaxCodeResponse.Rate was
  float64 → "cannot unmarshal string into float64". Add a flexFloat type that decodes number-or-string
  (re-marshals as a number). Fixes Settings -> Tax and per-item tax resolution.
- on-account 400 "credit sale requires a customer": when an order has no customer, attribute the
  credit to the tenant's shared marketflow "Walk-in" customer (best-effort UpsertContactByPhone on
  the +000000000000 sentinel) instead of failing. Inject the marketflow client into the payment service.

+++ internal/modules/orders/tax_resolver.go
@@ -75,7 +75,7 @@ func (r *TaxResolver) Resolve(ctx context.Context, tenantSlug, taxCodeID string)
 
 	info := &TaxCodeInfo{
 		Code:    tc.Code,
-		Rate:    tc.Rate,
+		Rate:    float64(tc.Rate),
 		KRACode: tc.KRACode,
 		TaxType: tc.TaxType,
 	}
